% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plist-package.R
\docType{package}
\name{plist}
\alias{plist}
\alias{plist-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/plist}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/plist/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
