% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plist_to_xml}
\alias{plist_to_xml}
\title{Convert a property list to raw (unparsed) XML}
\usage{
plist_to_xml(x)
}
\arguments{
\item{x}{binary or XML property list}
}
\description{
Convert a property list to raw (unparsed) XML
}
