% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_plist}
\alias{read_plist}
\title{Read a binary or XML property list from a file/connection}
\usage{
read_plist(x)
}
\arguments{
\item{x}{a file or connection (will be \code{\link[=path.expand]{path.expand()}}ed)}
}
\description{
Read a binary or XML property list from a file/connection
}
\examples{
read_plist(system.file("sample/plist.txt", package = "plist"))
read_plist(system.file("sample/plist.bin", package = "plist"))
}
