% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plist-package.R
\docType{package}
\name{plist}
\alias{plist}
\alias{plist-package}
\title{Read and Manipulate Binary or XML Property Lists}
\description{
A property list is a representation of a hierarchy of objects that
can be stored in the file system, memory, or databases and reconstituted later.
Property lists give applications a lightweight and portable way to store small
amounts of data. They are hierarchies of data made from specific types of
objects—they are, in effect, an object graph. Tools are provided to read
and manipulate binary or XML property lists.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/plist}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/plist/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
