% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plist_to_list}
\alias{plist_to_list}
\title{Convert a property list (binary or XML) to an R list}
\usage{
plist_to_list(x)
}
\arguments{
\item{x}{binary or XML property list}
}
\description{
Convert a property list (binary or XML) to an R list
}
