% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plist_to_xml}
\alias{plist_to_xml}
\title{Convert a property list to raw (unparsed) XML}
\usage{
plist_to_xml(x)
}
\arguments{
\item{x}{binary or XML property list}
}
\description{
Convert a property list to raw (unparsed) XML
}
\examples{
xml <- '<?xml version="1.0" encoding="UTF-8"?>
<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
<plist version="1.0">
<dict>
<key>Label</key>
<string>com.example.app</string>
<key>Program</key>
<string>/Users/Me/Scripts/cleanup.sh</string>
<key>RunAtLoad</key>
<true/>
</dict>
</plist>'

plist_to_xml(xml)
}
